backend/requirements.txt
easyocr
Flask
Pillow
